% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexfungorum.R
\name{ixf_AllUpdatedNames}
\alias{ixf_AllUpdatedNames}
\title{Return all Index Fungorum IDs modified since a certain date.}
\usage{
ixf_AllUpdatedNames(startDate, verbose = FALSE)
}
\arguments{
\item{startDate}{The starting date to search for updates. Should be a string
in \code{YYYYMMDD[HHMMSS]} format, or a \code{\link{Date}}, \code{\link{POSIXct}}, or
\code{\link{POSIXlt}} object.}

\item{verbose}{If \code{TRUE}, print a message when sending the query and
receiving the response.}
}
\value{
A character vector of numeric Index Fungorum record IDs.
}
\description{
The implementation currently use simple regex string processing on the query
result, rather than an XML parser, because this seems to be much, much
faster.  However, it may be vulnerable to a change in formatting by Index
Fungorum.
}
\examples{
ixf_AllUpdatedNames("20180101")
}
