% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexfungorum.R
\name{ixf_do_query}
\alias{ixf_do_query}
\title{Send a query to Index Fungorum, with exponential backoff on retries.}
\usage{
ixf_do_query(querytype, timeout = 60, maxtries = 10, tries = 0, wait = 1, ...)
}
\arguments{
\item{querytype}{type of query}

\item{timeout}{timeout for http request}

\item{maxtries}{maximum number of tries}

\item{tries}{number of tries that have already been made}

\item{wait}{seconds to wait between tries}

\item{...}{queries to send}
}
\value{
the query response
}
\description{
Send a query to Index Fungorum, with exponential backoff on retries.
}
