% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexfungorum.R
\name{ixf_NameByKey}
\alias{ixf_NameByKey}
\title{Look up the Index Fungorum record for a taxon.}
\usage{
ixf_NameByKey(NameKey, verbose = FALSE, schema = NULL, ...)
}
\arguments{
\item{NameKey}{The record number (as \code{integer} or \code{character}) of
the taxon. May have length >1, in which case multiple queries will be sent.}

\item{verbose}{If \code{TRUE}, print a message for each query sent to Index
Fungorum.}

\item{schema}{XML schema.  If the XML schema has already been parsed, it is
faster to send it rather than parse it again.}

\item{...}{extra arguments passed to \code{\link{ixf_do_query}}}
}
\value{
A \code{\link[tibble]{tibble}} with one row, containing the Index
Fungorum record for the taxon.
}
\description{
Look up the Index Fungorum record for a taxon.
}
\examples{
ixf_NameByKey(303114)
}
