% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexfungorum.R
\name{ixf_NameFullByKey}
\alias{ixf_NameFullByKey}
\title{Find the full name of a taxon on Index Fungorum}
\usage{
ixf_NameFullByKey(NameLsid, verbose = FALSE)
}
\arguments{
\item{NameLsid}{The record number (as \code{integer} or \code{character}) of
the taxon.  May have length >1, in which case multiple queries will be sent.}

\item{verbose}{If \code{TRUE}, print a message for each query sent to Index
Fungorum.}
}
\value{
The fully qualified scientific name of the taxon
}
\description{
Find the full name of a taxon on Index Fungorum
}
\examples{
ixf_NameFullByKey(303114)
}
