% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexfungorum.R
\name{ixf_NamesByCurrentKey}
\alias{ixf_NamesByCurrentKey}
\title{Look up all Index Fungorum records which list a given taxon as their current name.}
\usage{
ixf_NamesByCurrentKey(CurrentKey, verbose = FALSE, schema = NULL, ...)
}
\arguments{
\item{CurrentKey}{The record number (as \code{integer} or \code{character}) of
the current taxon. May have length >1, in which case multiple queries will be sent.}

\item{verbose}{If \code{TRUE}, print a message for each query sent to Index
Fungorum.}

\item{schema}{XML schema.  If the XML schema has already been parsed, it is
faster to send it rather than parse it again.}

\item{...}{extra arguments passed to \code{\link{ixf_do_query}}}
}
\value{
A \code{\link[tibble]{tibble}} with one row for each synonym,
containing the Index Fungorum records for the synonyms.
}
\description{
Look up all Index Fungorum records which list a given taxon as their current name.
}
\examples{
ixf_NameByCurrentKey(303114)
}
