% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexfungorum.R
\name{ixf_Search}
\alias{ixf_Search}
\alias{ixf_AuthorSearch}
\alias{ixf_NameSearch}
\alias{ixf_EpithetSearch}
\title{Search Index Fungorum}
\usage{
ixf_Search(SearchText, AnywhereInText = FALSE, MaxNumber = 1000,
  search_type = c("AuthorSearch", "EpithetSearch", "NameSearch"),
  verbose = FALSE, schema = NULL, ...)

ixf_AuthorSearch(SearchText, AnywhereInText = FALSE, MaxNumber = 1000,
  verbose = FALSE, schema = NULL, ...)

ixf_NameSearch(SearchText, AnywhereInText = FALSE, MaxNumber = 1000,
  verbose = FALSE, schema = NULL, ...)

ixf_EpithetSearch(SearchText, AnywhereInText = FALSE, MaxNumber = 1000,
  verbose = FALSE, schema = NULL, ...)
}
\arguments{
\item{SearchText}{Text to search for. May have length >1, in which case
multiple queries will be sent.}

\item{AnywhereInText}{If \code{FALSE}, \code{SearchText} must be the initial
substring. If \code{TRUE}, it can occur anywhere in the field}

\item{MaxNumber}{The maximum number of records to return}

\item{search_type}{The type of search to perform; i.e., which field will be searches.  Index Fungorum supports \code{"AuthorSearch"},
\code{"EpithetSearch"},
and \code{"NameSearch"}}

\item{verbose}{If \code{TRUE}, print a message for each query sent to Index
Fungorum.}

\item{schema}{for internal use.}

\item{...}{additional parameters to pass to \code{\link{ixf_do_query}}.  IN particular, \code{timeout} and \code{maxtries} may be useful.}
}
\value{
a \code{\link[tibble]{tibble}} containing Index Fungorum records.
}
\description{
Search Index Fungorum
}
\examples{
ixf_Search("muscaria", search_type = "Epithet")
ixf_AuthorSearch("Pers.", FALSE, 10)
ixf_NameSearch("Amanita muscaria")

ixf_EpithetSearch("cervesiae", FALSE, 10)
}
